+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/components/kill_gotcha.hpp
@@ -22,23 +22,38 @@
 
 #pragma once
 
-#include "common/defines.h"
+#include "core/defines.hpp"
 
-#define ROCPROFSYS_METADATA(...) ::tim::manager::add_metadata(__VA_ARGS__)
+#include <timemory/components/base.hpp>
+#include <timemory/components/gotcha/backends.hpp>
 
-#if !defined(ROCPROFSYS_DEFAULT_OBJECT)
-#    define ROCPROFSYS_DEFAULT_OBJECT(NAME)                                              \
-        NAME()                           = default;                                      \
-        NAME(const NAME&)                = default;                                      \
-        NAME(NAME&&) noexcept            = default;                                      \
-        NAME& operator=(const NAME&)     = default;                                      \
-        NAME& operator=(NAME&&) noexcept = default;
-#endif
+#include <csignal>
+#include <sys/types.h>
 
-#if !defined(ROCPROFSYS_DEFAULT_COPY_MOVE)
-#    define ROCPROFSYS_DEFAULT_COPY_MOVE(NAME)                                           \
-        NAME(const NAME&)                = default;                                      \
-        NAME(NAME&&) noexcept            = default;                                      \
-        NAME& operator=(const NAME&)     = default;                                      \
-        NAME& operator=(NAME&&) noexcept = default;
-#endif
+namespace rocprofsys
+{
+namespace component
+{
+struct kill_gotcha : tim::component::base<kill_gotcha, void>
+{
+    static constexpr size_t gotcha_capacity = 1;
+
+    using gotcha_data = tim::component::gotcha_data;
+    using kill_func_t = int (*)(pid_t, int);
+
+    ROCPROFSYS_DEFAULT_OBJECT(kill_gotcha)
+
+    static std::string label() { return "kill_gotcha"; }
+
+    static void configure();
+
+    static inline void start() {}
+    static inline void stop() {}
+
+    int operator()(const gotcha_data&, kill_func_t, pid_t, int) const;
+};
+}  // namespace component
+
+using kill_gotcha_t = tim::component::gotcha<component::kill_gotcha::gotcha_capacity,
+                                             std::tuple<>, component::kill_gotcha>;
+}  // namespace rocprofsys